move zlib back for josh
